# read_single_spectrum function to get speed up where desispec.io.read_spectra hits the I/O limits