The `desispec.io.read_spectra` function hits the I/O limitations of the CFS disks when reading spectra on multiple cores. This is because for lots of stacking purposes we are typically only interested in reading ~1 spectrum from each file we read in.

I have stripped back the `desispec.io.read_spectra` function to only retrieve the bare minimum data I need from just the spectrum I am interested in in each file. With this function the average read time of a spectrum is 8.45s based on a test of reading 10k random spectra on 128 cores (660s/10k spectra * 128 cores).

A speed up of ~3.6x compared to the original function with an average read time of 30.7s (1737s/7236 spectra * 128 cores).
However, it's still a significantly longer read time than the average read time when not hitting the max I/O read speeds where it's less than 1s per spectrum.